Add unique ownership to ChatBotPanelDialog object in chatgui.h
Updated the private member variable, `_chatLogic` in the `ChatBotPanelDialog` class to be a std::unique_ptr` instead of a raw pointer to force unique ownership rules.

This allowed for a cleaner construction without having to use `new` and also allowed us to use the default constructor instead of worrying about manually deallocating the memory in the heap where the variable lived.

**Other changes**
- Added `apt-get update` before the Python 3 PIP install in the Dockerfile since it was causing an error during the Docker build. See https://stackoverflow.com/questions/6587507/how-to-install-pip-with-python-3 for more info.