Allow MacOS Build through CMakeLists.txt to track down build error
Pulled in a local dependency via `brew install wxwidgets` to allow for local build and debugging.

Found this error during Docker build and again on local build:

```sh
Undefined symbols for architecture x86_64:
  "vtable for ChatBotPanelDialog", referenced from:
      ChatBotPanelDialog::ChatBotPanelDialog(wxWindow*, int) in chatgui.cpp.o
  NOTE: a missing vtable usually means the first non-inline virtual member function has no definition.
ld: symbol(s) not found for architecture x86_64
```

Found out it was from getting rid of the destructor for `ChatPanelDialog`. Reverted that change.